Moving tests into existing file.
